Add witting tei to file-like object
Convention in Python APIs is usually either a string or a file-like
object in the same argument. This would be a suitable alternative but
would either involve API incompatibility: argument renamed from (filename
to file) or confusing naming when leaving the name as is.

The current implementation is already a bit weird in that it can write
to both stdout and a file (by filename), now it can additionally (in one
single call) also write to a file-like object.